Move LineRange to Utility namespace.
